app/config/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via
    #   django
    #   scout-apm
astroid==2.15.4
    # via pylint
asttokens==2.2.1
    # via stack-data
async-timeout==4.0.2
    # via redis
backcall==0.2.0
    # via ipython
black==23.3.0
    # via -r requirements.in
boto3==1.26.123
    # via
    #   -r requirements.in
    #   django-ses
botocore==1.29.123
    # via
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via whitenoise
build==0.10.0
    # via pip-tools
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
    #   urllib3
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
coverage==7.2.4
    # via -r requirements.in
cryptography==40.0.2
    # via
    #   django-ses
    #   pyjwt
    #   pyopenssl
    #   urllib3
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
defusedxml==0.7.1
    # via python3-openid
dill==0.3.6
    # via pylint
dj-database-url==2.0.0
    # via -r requirements.in
django==4.1.7
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-allauth
    #   django-browser-reload
    #   django-clearcache
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-embed-video
    #   django-lifecycle
    #   django-markdownx
    #   django-redis
    #   django-ses
    #   django-storages
django-allauth==0.54.0
    # via -r requirements.in
django-browser-reload==1.8.0
    # via -r requirements.in
django-clearcache==1.2.1
    # via -r requirements.in
django-cors-headers==3.14.0
    # via -r requirements.in
django-debug-toolbar==4.0.0
    # via -r requirements.in
django-embed-video==1.4.8
    # via -r requirements.in
django-lifecycle==1.0.0
    # via -r requirements.in
django-markdownx==4.0.2
    # via -r requirements.in
django-redis==5.2.0
    # via -r requirements.in
django-ses[bounce]==3.4.1
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via -r requirements.in
faker==18.6.0
    # via factory-boy
flake8==6.0.0
    # via flake8-isort
flake8-isort==6.0.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via
    #   requests
    #   urllib3
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via -r requirements.in
ipython==8.13.1
    # via
    #   -r requirements.in
    #   ipdb
isort==5.12.0
    # via
    #   flake8-isort
    #   pylint
jedi==0.18.2
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lazy-object-proxy==1.9.0
    # via astroid
markdown==3.4.3
    # via
    #   -r requirements.in
    #   django-markdownx
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mypy-extensions==1.0.0
    # via black
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via
    #   black
    #   build
    #   django-lifecycle
    #   pytest
    #   pytest-sugar
parso==0.8.3
    # via jedi
pathspec==0.11.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.5.0
    # via
    #   -r requirements.in
    #   django-markdownx
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.5.0
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.38
    # via ipython
psutil==5.9.5
    # via scout-apm
psycopg2-binary==2.9.6
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via flake8
pygments==2.15.1
    # via ipython
pyjwt[crypto]==2.6.0
    # via django-allauth
pylint==2.17.3
    # via
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.3
    # via -r requirements.in
pylint-plugin-utils==0.7
    # via pylint-django
pyopenssl==23.1.1
    # via urllib3
pyproject-hooks==1.0.0
    # via build
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-sugar
pytest-django==4.5.2
    # via -r requirements.in
pytest-sugar==0.9.7
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via django-ses
pywatchman==1.4.1
    # via -r requirements.in
redis==4.5.4
    # via django-redis
requests==2.29.0
    # via
    #   django-allauth
    #   django-embed-video
    #   django-ses
    #   requests-oauthlib
    #   stripe
requests-oauthlib==1.3.1
    # via django-allauth
s3transfer==0.6.0
    # via boto3
scout-apm==2.26.1
    # via -r requirements.in
sentry-sdk==1.21.1
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.2
    # via ipython
stripe==5.4.0
    # via -r requirements.in
termcolor==2.3.0
    # via pytest-sugar
tomlkit==0.11.8
    # via pylint
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.5.0
    # via dj-database-url
urllib3[secure]==1.26.15
    # via
    #   botocore
    #   requests
    #   scout-apm
    #   sentry-sdk
urllib3-secure-extra==0.1.0
    # via urllib3
urlman==2.0.1
    # via django-lifecycle
wcwidth==0.2.6
    # via prompt-toolkit
wheel==0.40.0
    # via pip-tools
whitenoise[brotli]==6.4.0
    # via -r requirements.in
wrapt==1.15.0
    # via
    #   astroid
    #   scout-apm

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
